Fixes and readability improvements in code
Fixed the problem where if there was only one hyperlink connected the link did not work and when all the hyperlinks were connected it worked as if there was only one connected. Also spaced out the paragraphs in the code so that there is no need to scroll along the editor to read them.